requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.0.0,<3.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.0
jinja2==3.1.2
google-generativeai==0.3.1
reportlab==4.0.4
weasyprint==60.1
xhtml2pdf==0.2.11
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
